+++ Braintree/Properties/AssemblyInfo.cs
@@ -29,5 +29,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("2.55.0.0")]
-[assembly: AssemblyFileVersion("2.55.0.0")]
+[assembly: AssemblyVersion("2.56.0.0")]
+[assembly: AssemblyFileVersion("2.56.0.0")]